ci: sadly reverting to no android device specific tests' -s
The only way to do this is using Firebase Test Lab, but there I doubt we can run the cafes. It's also possible on a macOS VM, but also a nightmare and expensive for tests that would run so frequently. Perhaps we could enable this for master builds in the future.

Signed-off-by: Sander Pick <sanderpick@gmail.com>